+++ pkg/blockbuilder/tsdb.go
@@ -33,14 +33,16 @@ import (
 )
 
 type TSDBBuilder struct {
-	logger log.Logger
+	dataDir string
+
+	logger           log.Logger
+	limits           *validation.Overrides
+	blocksStorageCfg mimir_tsdb.BlocksStorageConfig
+	metrics          tsdbBuilderMetrics
 
 	// Map of a tenant in a partition to its TSDB.
 	tsdbsMu sync.RWMutex
 	tsdbs   map[tsdbTenant]*userTSDB
-
-	limits *validation.Overrides
-	config mimir_tsdb.BlocksStorageConfig
 }
 
 type tsdbTenant struct {
@@ -48,12 +50,14 @@ type tsdbTenant struct {
 	tenantID    string
 }
 
-func NewTSDBBuilder(logger log.Logger, limits *validation.Overrides, config mimir_tsdb.BlocksStorageConfig) *TSDBBuilder {
+func NewTSDBBuilder(logger log.Logger, dataDir string, blocksStorageCfg mimir_tsdb.BlocksStorageConfig, limits *validation.Overrides, metrics tsdbBuilderMetrics) *TSDBBuilder {
 	return &TSDBBuilder{
-		tsdbs:  make(map[tsdbTenant]*userTSDB),
-		logger: logger,
-		limits: limits,
-		config: config,
+		dataDir:          dataDir,
+		logger:           logger,
+		limits:           limits,
+		blocksStorageCfg: blocksStorageCfg,
+		metrics:          metrics,
+		tsdbs:            make(map[tsdbTenant]*userTSDB),
 	}
 }
 
@@ -62,8 +66,8 @@ func NewTSDBBuilder(logger log.Logger, limits *validation.Overrides, config mimi
 // where the sample was not put in the TSDB because it was discarded or was already processed before.
 // lastBlockMax: max time of the block in the previous block building cycle.
 // blockMax: max time of the block in the current block building cycle. This blockMax is exclusive of the last sample by design in TSDB.
-// recordProcessedBefore: true if the record was processed in the previous cycle. (It gets processed again if some samples did not fit in the previous cycle.)
-func (b *TSDBBuilder) Process(ctx context.Context, rec *kgo.Record, lastBlockMax, blockMax int64, recordProcessedBefore bool) (_ bool, err error) {
+// recordAlreadyProcessed: true if the record was processed in the previous cycle. (It gets processed again if some samples did not fit in the previous cycle.)
+func (b *TSDBBuilder) Process(ctx context.Context, rec *kgo.Record, lastBlockMax, blockMax int64, recordAlreadyProcessed bool) (_ bool, err error) {
 	userID := string(rec.Key)
 
 	req := mimirpb.PreallocWriteRequest{
@@ -119,7 +123,7 @@ func (b *TSDBBuilder) Process(ctx context.Context, rec *kgo.Record, lastBlockMax
 				allSamplesProcessed = false
 				continue
 			}
-			if recordProcessedBefore && s.TimestampMs < lastBlockMax {
+			if recordAlreadyProcessed && s.TimestampMs < lastBlockMax {
 				// This sample was already processed in the previous cycle.
 				continue
 			}
@@ -150,7 +154,7 @@ func (b *TSDBBuilder) Process(ctx context.Context, rec *kgo.Record, lastBlockMax
 				allSamplesProcessed = false
 				continue
 			}
-			if recordProcessedBefore && h.Timestamp < lastBlockMax {
+			if recordAlreadyProcessed && h.Timestamp < lastBlockMax {
 				// This sample was already processed in the previous cycle.
 				continue
 			}
@@ -260,7 +264,7 @@ func (b *TSDBBuilder) getOrCreateTSDB(tenant tsdbTenant) (*userTSDB, error) {
 }
 
 func (b *TSDBBuilder) newTSDB(tenant tsdbTenant) (*userTSDB, error) {
-	udir := filepath.Join(b.config.TSDB.Dir, strconv.Itoa(int(tenant.partitionID)), tenant.tenantID)
+	udir := filepath.Join(b.dataDir, strconv.Itoa(int(tenant.partitionID)), tenant.tenantID)
 	// Remove any previous TSDB dir. We don't need it.
 	if err := os.RemoveAll(udir); err != nil {
 		return nil, err
@@ -281,16 +285,16 @@ func (b *TSDBBuilder) newTSDB(tenant tsdbTenant) (*userTSDB, error) {
 		MinBlockDuration:            2 * time.Hour.Milliseconds(),
 		MaxBlockDuration:            2 * time.Hour.Milliseconds(),
 		NoLockfile:                  true,
-		StripeSize:                  b.config.TSDB.StripeSize,
-		HeadChunksWriteBufferSize:   b.config.TSDB.HeadChunksWriteBufferSize,
-		HeadChunksWriteQueueSize:    b.config.TSDB.HeadChunksWriteQueueSize,
+		StripeSize:                  b.blocksStorageCfg.TSDB.StripeSize,
+		HeadChunksWriteBufferSize:   b.blocksStorageCfg.TSDB.HeadChunksWriteBufferSize,
+		HeadChunksWriteQueueSize:    b.blocksStorageCfg.TSDB.HeadChunksWriteQueueSize,
 		WALSegmentSize:              -1, // No WAL
 		SeriesLifecycleCallback:     udb,
 		BlocksToDelete:              udb.blocksToDelete,
 		IsolationDisabled:           true,
 		EnableOverlappingCompaction: false,                                                // always false since Mimir only uploads lvl 1 compacted blocks
 		OutOfOrderTimeWindow:        b.limits.OutOfOrderTimeWindow(userID).Milliseconds(), // The unit must be same as our timestamps.
-		OutOfOrderCapMax:            int64(b.config.TSDB.OutOfOrderCapacityMax),
+		OutOfOrderCapMax:            int64(b.blocksStorageCfg.TSDB.OutOfOrderCapacityMax),
 		EnableNativeHistograms:      b.limits.NativeHistogramsIngestionEnabled(userID),
 		SecondaryHashFunction:       nil, // TODO(codesome): May needed when applying limits. Used to determine the owned series by an ingesters
 	}, nil)
@@ -335,9 +339,8 @@ func (b *TSDBBuilder) CompactAndUpload(ctx context.Context, uploadBlocks blockUp
 
 		err = merr.Err()
 
-		// Clear the map so that it can be released from the memory. Not setting to nil in case
-		// we want to reuse the TSDBBuilder.
-		b.tsdbs = make(map[tsdbTenant]*userTSDB)
+		// Clear the map so that it can be released from the memory. Not setting to nil in case we want to reuse the TSDBBuilder.
+		clear(b.tsdbs)
 	}()
 
 	level.Info(b.logger).Log("msg", "compacting and uploading blocks", "num_tsdb", len(b.tsdbs))
@@ -349,12 +352,24 @@ func (b *TSDBBuilder) CompactAndUpload(ctx context.Context, uploadBlocks blockUp
 	numBlocks := atomic.NewInt64(0)
 
 	eg, ctx := errgroup.WithContext(ctx)
-	if b.config.TSDB.ShipConcurrency > 0 {
-		eg.SetLimit(b.config.TSDB.ShipConcurrency)
+	if b.blocksStorageCfg.TSDB.ShipConcurrency > 0 {
+		eg.SetLimit(b.blocksStorageCfg.TSDB.ShipConcurrency)
 	}
 	for tenant, db := range b.tsdbs {
-		eg.Go(func() error {
-			// TODO(codesome): add a metric for compaction and upload failures. An alert on this will be useful.
+		eg.Go(func() (err error) {
+			defer func(t time.Time) {
+				if errors.Is(err, context.Canceled) {
+					// Don't track any metrics if context was cancelled. Otherwise, it might be misleading.
+					return
+				}
+				partitionStr := fmt.Sprintf("%d", tenant.partitionID)
+				if err != nil {
+					b.metrics.compactAndUploadFailed.WithLabelValues(partitionStr).Inc()
+					return
+				}
+				b.metrics.compactAndUploadDuration.WithLabelValues(partitionStr).Observe(time.Since(t).Seconds())
+			}(time.Now())
+
 			if err := db.compactEverything(ctx); err != nil {
 				return err
 			}
@@ -414,11 +429,6 @@ type extendedAppender interface {
 type userTSDB struct {
 	*tsdb.DB
 	userID string
-}
-
-// BlocksUploader interface is used to have an easy way to mock it in tests.
-type BlocksUploader interface {
-	Sync(ctx context.Context) (uploaded int, err error)
 }
 
 func (u *userTSDB) compactEverything(ctx context.Context) error {